+++ internal/future/app/backend/backend_test.go
@@ -19,12 +19,12 @@ import (
 
 	"google.golang.org/grpc"
 	"open-match.dev/open-match/internal/future/pb"
-	"open-match.dev/open-match/internal/future/serving"
+	"open-match.dev/open-match/internal/future/serving/rpc"
 	servingTesting "open-match.dev/open-match/internal/future/serving/testing"
 )
 
 func TestServerBinding(t *testing.T) {
-	bs := func(p *serving.Params) {
+	bs := func(p *rpc.Params) {
 		p.AddHandleFunc(func(s *grpc.Server) {
 			pb.RegisterBackendServer(s, &backendService{})
 		}, pb.RegisterBackendHandlerFromEndpoint)
