+++ internal/future/app/mmlogic/mmlogic.go
@@ -19,7 +19,7 @@ import (
 	"google.golang.org/grpc"
 	"open-match.dev/open-match/internal/config"
 	"open-match.dev/open-match/internal/future/pb"
-	"open-match.dev/open-match/internal/future/serving"
+	"open-match.dev/open-match/internal/future/serving/rpc"
 )
 
 var (
@@ -37,7 +37,7 @@ func RunApplication() {
 			"error": err.Error(),
 		}).Fatalf("cannot read configuration.")
 	}
-	p, err := serving.NewParamsFromConfig(cfg, "api.mmlogic")
+	p, err := rpc.NewParamsFromConfig(cfg, "api.mmlogic")
 	if err != nil {
 		mmlogicLogger.WithFields(logrus.Fields{
 			"error": err.Error(),
@@ -50,11 +50,11 @@ func RunApplication() {
 		}).Fatalf("failed to bind mmlogic service.")
 	}
 
-	serving.MustServeForever(p)
+	rpc.MustServeForever(p)
 }
 
 // BindService creates the mmlogic service and binds it to the serving harness.
-func BindService(p *serving.Params, cfg config.View) error {
+func BindService(p *rpc.Params, cfg config.View) error {
 	service, err := newMmlogic(cfg)
 	if err != nil {
 		return err
